requirements.txt
aioboto3==10.1.0
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
blis==0.7.8
boto3==1.24.59
botocore==1.27.59
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.0.4
constantly==15.1.0
cryptography==38.0.1
cymem==2.0.6
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
murmurhash==1.0.8
numpy==1.23.3
packaging==21.3
pathy==0.6.2
preshed==3.0.7
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.4
supervisor==4.2.4
thinc==8.0.17
tqdm==4.64.1
Twisted==22.8.0
txaio==22.2.1
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.9.0
zope.interface==5.4.0
